+++ src/Plugins/Module/Formation/src/FormationWidget.cpp
@@ -56,6 +56,7 @@
 #include <Model/AircraftInfo.h>
 #include <Model/Position.h>
 #include <Model/PositionData.h>
+#include <Persistence/LogbookManager.h>
 #include <Persistence/Service/FlightService.h>
 #include <Persistence/Service/AircraftService.h>
 #include <PluginManager/SkyConnectManager.h>
@@ -270,6 +271,11 @@ void FormationWidget::initTimeOffsetUi() noexcept
 
 void FormationWidget::frenchConnection() noexcept
 {
+    // Logbook
+    const Logbook &logbook = Logbook::getInstance();
+    connect(&LogbookManager::getInstance(), &LogbookManager::connectionChanged,
+            this, &FormationWidget::updateUi);
+
     // Flight
     Flight &flight = Logbook::getInstance().getCurrentFlight();
     connect(&flight, &Flight::userAircraftChanged,